refactor(std/example): Inconsistencies in the example tests (#7684)
