+++ std/examples/catj_test.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { assertStrictEquals } from "../../testing/asserts.ts";
-import { resolve, dirname, fromFileUrl } from "../../path/mod.ts";
+import { assertStrictEquals } from "../testing/asserts.ts";
+import { dirname, fromFileUrl } from "../path/mod.ts";
 
-const moduleDir = resolve(dirname(fromFileUrl(import.meta.url)), "..");
+const moduleDir = dirname(fromFileUrl(import.meta.url));
 
 Deno.test("[examples/catj] print an array", async () => {
   const decoder = new TextDecoder();
