+++ std/examples/echo_server_test.ts
@@ -1,9 +1,9 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { assertStrictEquals, assertNotEquals } from "../../testing/asserts.ts";
-import { BufReader, ReadLineResult } from "../../io/bufio.ts";
-import { resolve, dirname, fromFileUrl } from "../../path/mod.ts";
+import { assertStrictEquals, assertNotEquals } from "../testing/asserts.ts";
+import { BufReader, ReadLineResult } from "../io/bufio.ts";
+import { dirname, fromFileUrl } from "../path/mod.ts";
 
-const moduleDir = resolve(dirname(fromFileUrl(import.meta.url)), "..");
+const moduleDir = dirname(fromFileUrl(import.meta.url));
 
 Deno.test("[examples/echo_server]", async () => {
   const encoder = new TextEncoder();
