+++ std/examples/test_test.ts
@@ -1,2 +1,2 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import "../test.ts";
+import "./test.ts";
